cleanup /W4 warnings on windows
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21167 f0d54148-0727-0410-94bb-9a71ac55c965

+++ alg/gdaltransformer.cpp
@@ -349,7 +349,7 @@ GDALSuggestedWarpOutput2( GDALDatasetH hSrcDS,
     int    nInYSize = GDALGetRasterYSize( hSrcDS );
 
 #define N_PIXELSTEP 50
-    int nSteps = double(MIN(nInYSize, nInXSize)) / N_PIXELSTEP + .5;
+    int nSteps = (int) (double(MIN(nInYSize, nInXSize)) / N_PIXELSTEP + .5);
     if (nSteps < 20)
         nSteps = 20;
